deplacement semi OK, turn OK

+++ header.h
@@ -109,7 +109,7 @@ struct				s_player
 	int		y;
 	int		z;
 	float	speed;
-	float	sensi;
+	double	sensi;
 	//int		height;
 	double	fieldvis; //le champ de vision en radian
 	double	angle; //l'angle en radian par rapport a l'axe des abscisses
@@ -193,7 +193,7 @@ bool				raycastx(t_wall *wall, t_polygon polygon, t_data data);
 void				renderbsp(t_data *data, t_node current, t_player player);
 void				replace_poly(t_polygon *polygon, t_player player);
 void				reset_data(t_data *data);
-void				set_player_angle(t_player *player, float angle);
+void				set_player_angle(t_player *player, double angle);
 void				set_player_pos(t_player *player, float x, float z);
 void				set_cub(t_cub *cub, int i, int counter);
 void				set_delta(t_wall *wall);
